apps/api/requirements-ai.txt
# AI/ML Dependencies for "AI Property Doctor"
# Advanced computer vision, forecasting, and NLP

# Core ML frameworks
xgboost==2.1.3
lightgbm==4.5.0
catboost==1.2.7

# Time series forecasting
prophet==1.1.6
statsmodels==0.14.4

# Computer Vision (for property analysis from photos)
torch==2.5.1
torchvision==0.20.1
opencv-python==4.10.0.84
pillow==11.0.0
ultralytics==8.3.41  # YOLO for object detection

# French NLP (CamemBERT)
transformers==4.47.1
sentencepiece==0.2.0
tokenizers==0.21.0

# Image generation (renovation renders)
diffusers==0.32.1
accelerate==1.2.1

# Geospatial analysis
geopandas==1.0.1
shapely==2.0.6
folium==0.18.0

# Scientific computing
numpy==2.1.3
pandas==2.2.3
scipy==1.14.1
scikit-learn==1.6.0

# Visualization
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1

# PDF generation (for reports)
reportlab==4.2.5
pypdf2==3.0.1
